refactor : 컴포넌트 파일 경로 수정
